Change some dev ergonomics
This sets up the webapp to run in a "dev" mode without uvicorn in a
local dev environment with the ability to run it in a "prod" mode if
needed.

This adjusts the Makefile to use docker compose things because the
webapp requires the otel-collector now.

This adds some basic documentation to the README.

This moves entrypoint and webapp run bits to separate scripts. This
makes it easier to move things to environment variables later.

This adds a .dockerignore so that the .git directory isn't added to the
docker image.