Get linting and tests working
This fixes configuration so the otel-collector service is running and
the web application is configured correctly to send opentelemetry trace
and metrics data to it.

This gets ruff linting working.

This moves all the opentelemetry related code into
`app/observability.py` so it's all in one place.

This adds a basic test for the `/` endpoint asserting things about the
response payload.

This adds environment configuration documentation to the `README.md`.